requirements.txt
apache-beam==2.48.0
apache-flink==1.20.0
apache-flink-libraries==1.20.0
avro-python3==1.10.2
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.2.1
contourpy==1.3.1
crcmod==1.7
cycler==0.12.1
Cython==3.0.12
dill==0.3.1.1
dnspython==2.7.0
docopt==0.6.2
fastavro==1.10.0
fasteners==0.19
find_libpython==0.4.0
Flask==3.1.0
fonttools==4.56.0
grpcio==1.70.0
hdfs==2.7.3
httplib2==0.22.0
idna==3.10
influxdb-client==1.48.0
influxdb3-python==0.11.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kafka-python==2.0.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==1.24.4
objsize==0.6.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pemja==0.4.1
pillow==11.1.0
pmdarima==2.0.4
proto-plus==1.26.0
protobuf==4.23.4
py4j==0.10.9.7
pyarrow==11.0.0
pydot==1.4.2
pymongo==4.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
reactivex==4.0.4
regex==2024.11.6
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
statsmodels==0.14.4
tabulate==0.9.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
zstandard==0.23.0
